Fixes inconsistent behavior of RateLimitStatus.
The result between createFromJSONResponse and createFromResponseHeader differed.
createFromJSONResponse initialized resetTimeInSeconds with the same value as resetTime.getTime()/1000 while createFromResponseHeader initialized it with the amount of seconds until the next reset.

I changed the code to initialize resetTimeInSeconds consistently with resetTime.getTime()/1000 and added another attribute, secondsUntilReset, that is initialized consistently with the amount of seconds until the next reset.

I'd usually remove resetTimeInSeconds since it just duplicates the Date but I'll leave that to you.